requirements.txt
django
django-cors-headers 
djangorestframework
django-environ
ffmpeg-python
psycopg2
faster-whisper
